service_bot/service_bot/requirements.txt
aiogram==3.2.0
python-dotenv>=1.0.0
aiohttp>=3.8.5
pydantic>=2.0.0
loguru>=0.7.0
